requirements.txt
requests
python-dotenv
apache-airflow
pandas

pdfplumber
openai
pymongo
psycopg2-binary